first part of socket code shuffle
